api_yamdb/requirements.txt
requests==2.26.0
django==2.2.16
djangorestframework==3.12.4
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.20.0
djangorestframework-simplejwt==5.1.0
django-filter==21.1
gunicorn==20.0.4
psycopg2-binary==2.8.6